factored out the node class
