+++ src/lib/mapCreator.h
@@ -9,50 +9,10 @@
 #include "components/Renderable.hpp"
 #include "components/Transform.hpp"
 #include "core/Enums.hpp"
+#include "mapGeneration/BinarySpacePartitioner.h"
+#include "mapGeneration/Node.h"
 
 const int MAP_TALLNESS = 25;
-
-class Node;
-typedef std::shared_ptr<Node> NodePtr;
-typedef std::weak_ptr<Node> WeakNodePtr;
-
-class Node : public std::enable_shared_from_this<Node>, public Transform, public Renderable {
- public:
-  std::vector<NodePtr> children;
-  vec2 bottomLeftCorner, topLeftCorner, bottomRightCorner, topRightCorner;
-  int treeLayerIndex;
-  WeakNodePtr parent;
-
-  Node() : treeLayerIndex(0){};  // default constructor
-  Node(const vec2& bottomLeft, const vec2& topRight, int index);
-
-  void addChild(std::shared_ptr<Node> child);
-  int getWidth();
-  int getHeight();
-};
-
-enum Orientation { HORIZONTAL, VERTICAL, NONE };
-
-class Line {
- public:
-  Orientation orientation;
-  vec2 coordinates;
-  Line(Orientation orientation, vec2 coordinates);
-};
-
-class BinarySpacePartitioner {
- public:
-  NodePtr rootNode;
-
-  BinarySpacePartitioner(int mapWidth, int mapHeight);
-
-  vec2 getCoordinatesForOrientation(Orientation orientation, vec2 bottomLeftCorner, vec2 topRightCorner,
-                                    int minRoomWidth, int minRoomHeight);
-  Line GetLineDividingSpace(vec2 bottomLeftCorner, vec2 topRightCorner, int minRoomWidth, int minRoomHeight);
-  void splitTheSpace(NodePtr currentNode, int minRoomWidth, int minRoomHeight);
-  std::vector<NodePtr> collectAllNodesIteratively(NodePtr inputRootNode);
-  std::vector<NodePtr> prepareNodesCollection(int maxIterations, int minRoomWidth, int minRoomHeight);
-};
 
 std::vector<NodePtr> findLeafNodes(const NodePtr& rootNode);
